requirements.txt
langserve[server]
fastapi
langchain
uvicorn
langchain_core
langchain_openai
requests
python-dotenv